pyproject.toml
[tool.poetry]
name = "learn-sql"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]
license = "Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
hypothesis = {extras = ["cli"], version = "^6.87.1"}
mysql-connector-python = "^8.1.0"
numpy = "^1.26.0"
pandas = "^2.1.1"
psycopg2-binary = "2.9.6"
pymysql = "^1.1.0"
python-decouple = "^3.8"
sqlalchemy = "2.0.13"

[tool.poetry.group.dev.dependencies]
ipython = "^8.4.0"
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
pytest-cov = "^4.0.0"
pytest-datafiles = "^3.0.0"
pytest-xdist = "^2.5.0"
rich = "^12.6.0"
ruff = "^0.0.269"
hypothesis = "^6.87.1"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []
# Skip unused variable rules
extend-ignore = ["D203", "E203", "E251", "E266", "E401", "E402", "E501", "F401", "F403"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    "dist",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    "__pycache__",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Black (default: 88)
line-length = 130

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
